Nanopowers

+++ src/main.cpp
@@ -9,30 +9,69 @@
 #include "NPCManager.hpp"
 #include "TransportManager.hpp"
 #include "Database.hpp"
+#include "TableData.hpp"
 
 #include "settings.hpp"
 
 #if defined(__MINGW32__) && !defined(_GLIBCXX_HAS_GTHREADS)
     #include "mingw/mingw.thread.h"
-#else 
+#else
 #include <thread>
 #endif
 #include <string>
+#include <signal.h>
 
-CNShardServer *shardServer;
-std::thread *shardThread;
+// HACK
+#ifdef __has_feature
+#if __has_feature(address_sanitizer)
+#define __SANITIZE_ADDRESS__ 1
+#endif
+#endif
+
+CNShardServer *shardServer = nullptr;
+std::thread *shardThread = nullptr;
 
 void startShard(CNShardServer* server) {
     server->start();
 }
 
+#ifndef _WIN32
 // terminate gracefully on SIGINT (for gprof)
 void terminate(int arg) {
-    std::cout << "OpenFusion: terminating" << std::endl;
-    shardServer->kill();
-    shardThread->join();
+    std::cout << "OpenFusion: terminating." << std::endl;
+
+    if (shardServer != nullptr && shardThread != nullptr) {
+        shardServer->kill();
+        shardThread->join();
+    }
+
+#if defined(__SANITIZE_ADDRESS__)
+    TableData::cleanup();
+#endif
+
     exit(0);
 }
+
+void initsignals() {
+    struct sigaction act;
+
+    memset((void*)&act, 0, sizeof(act));
+    sigemptyset(&act.sa_mask);
+
+    // tell the OS to not kill us if you use a broken pipe, just let us know thru recv() or send()
+    act.sa_handler = SIG_IGN;
+    if (sigaction(SIGPIPE, &act, NULL) < 0) {
+        perror("sigaction");
+        exit(1);
+    }
+
+    act.sa_handler = terminate;
+    if (sigaction(SIGINT, &act, NULL) < 0) {
+        perror("sigaction");
+        exit(1);
+    }
+}
+#endif
 
 int main() {
 #ifdef _WIN32
@@ -42,13 +81,12 @@ int main() {
         exit(EXIT_FAILURE);
     }
 #else
-    // tell the OS to not kill us if you use a broken pipe, just let us know thru recv() or send()
-    signal(SIGPIPE, SIG_IGN);
-    signal(SIGINT, terminate); // TODO: use sigaction() instead
+    initsignals();
 #endif
     settings::init();
     std::cout << "[INFO] Protocol version: " << PROTOCOL_VERSION << std::endl;
     std::cout << "[INFO] Intializing Packet Managers..." << std::endl;
+    TableData::init();
     PlayerManager::init();
     ChatManager::init();
     CombatManager::init();
